requirements.txt
gTTS==2.5.4
pillow==10.2.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
pydub
moviepy==1.0.3

